VKT(Frontend): Refactor: load initial data on PublicHomePage and ClerkHomePage like with other projects
