+++ frontend/packages/vkt/src/components/publicExamEvent/PublicExamEventGrid.tsx
@@ -1,6 +1,5 @@
 import OpenInNewIcon from '@mui/icons-material/OpenInNew';
 import { Alert, Grid, Paper } from '@mui/material';
-import { useEffect } from 'react';
 import { Trans } from 'react-i18next';
 import { ExtLink, H1, H2, HeaderSeparator, Text } from 'shared/components';
 import { APIResponseStatus, Severity } from 'shared/enums';
@@ -8,8 +7,7 @@ import { APIResponseStatus, Severity } from 'shared/enums';
 import { PublicExamEventListing } from 'components/publicExamEvent/listing/PublicExamEventListing';
 import { PublicExamEventGridSkeleton } from 'components/skeletons/PublicExamEventGridSkeleton';
 import { useCommonTranslation, usePublicTranslation } from 'configs/i18n';
-import { useAppDispatch, useAppSelector } from 'configs/redux';
-import { loadPublicExamEvents } from 'redux/reducers/publicExamEvent';
+import { useAppSelector } from 'configs/redux';
 import { publicExamEventsSelector } from 'redux/selectors/publicExamEvent';
 
 export const PublicExamEventGrid = () => {
@@ -19,17 +17,10 @@ export const PublicExamEventGrid = () => {
 
   // Redux
   const { status, examEvents } = useAppSelector(publicExamEventsSelector);
-  const dispatch = useAppDispatch();
 
   // State
   const isLoading = status === APIResponseStatus.InProgress;
   const hasResults = examEvents.length > 0;
-
-  useEffect(() => {
-    if (status === APIResponseStatus.NotStarted) {
-      dispatch(loadPublicExamEvents());
-    }
-  }, [dispatch, status]);
 
   return (
     <>
